Refactor Automapper. Adeded jenkins pipline
